added script error location tracking for runtime errors.
every expression now gets assigned an origin. For parsed objects, this is an ExprFromSourceOrigin that contains filename, column, and line info.
For expanded macros, an ExprFromMacroOrigin points to the expanded expression, and hence its (probably source code) origin.
For rewritten expressions, an ExprFromRewriteOrigin points to the rewritten expression, and also stores the name of the rewrite.

The core interpreter evaluation methods eval, iter (from Expr) and moveNext (from JsonIterator) now wrap every exception
into an ExprEvalException that points to the Expr object causing the error. The toString method prints this error location. To avoid every
expression to duplicate this code, theese methods are now final, and call corresponding ...Raw() methods. These are the new method names to override in derived expressions or iterators.

To propagate error locations to remote instances of the interpreter (Map/Reduce), the parser (actually the lexer) understands a #() directive that
is similar to the C/C++ preprocessor directive #line, and allows to set the reported error location in a source file.
The decompile method in Expr now has an addditional boolean parameter flag. If set, it emits #() directives based on the Origin objects of the expressions.
Currently, this is only set to true for the decompiled job descriptions used in MapReduceBase. To avoid cluttering the decompile code for every expression type, this is done in the root decompile method, which is now final, and calls decompileRaw, which should be overriden in those expressiosn that require special decompilation.